refactor(function): move soundex module
