--- src/query/functions-v2/src/scalars/soundex.rs
@@ -1,44 +1,0 @@
-// Copyright 2021 Datafuse Labs.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-pub(crate) struct Soundex;
-
-impl Soundex {
-    #[inline(always)]
-    pub fn number_map(i: char) -> Option<u8> {
-        match i.to_ascii_lowercase() {
-            'b' | 'f' | 'p' | 'v' => Some(b'1'),
-            'c' | 'g' | 'j' | 'k' | 'q' | 's' | 'x' | 'z' => Some(b'2'),
-            'd' | 't' => Some(b'3'),
-            'l' => Some(b'4'),
-            'm' | 'n' => Some(b'5'),
-            'r' => Some(b'6'),
-            _ => Some(b'0'),
-        }
-    }
-
-    #[inline(always)]
-    pub fn is_drop(c: char) -> bool {
-        matches!(
-            c.to_ascii_lowercase(),
-            'a' | 'e' | 'i' | 'o' | 'u' | 'y' | 'h' | 'w'
-        )
-    }
-
-    // https://github.com/mysql/mysql-server/blob/3290a66c89eb1625a7058e0ef732432b6952b435/sql/item_strfunc.cc#L1919
-    #[inline(always)]
-    pub fn is_uni_alphabetic(c: char) -> bool {
-        ('a'..='z').contains(&c) || ('A'..='Z').contains(&c) || c as i32 >= 0xC0
-    }
-}
